requirements.txt
# Core Dependencies
telethon>=1.34.0
python-dotenv>=1.0.0

# Web Dashboard Dependencies
flask>=2.0.0
flask-socketio>=5.0.0
python-socketio>=5.0.0

# Data Processing
numpy>=1.21.0

# For API_ID and API_HASH:
# 1. Visit https://my.telegram.org
# 2. Login with your account
# 3. Go to "API development tools"
# 4. Create a new application
# 5. Copy API_ID and API_HASH